backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
python-dotenv
python-multipart
pymupdf
pillow
pydantic
httpx
pytest
